feat: move menu positioning logic to separate file
